Changing regular expression to allow @ symbols in package names
Some npm packages have @ symbols at the beginning, and tasklemon's regexes that parse package+version combos mistake the initial @
for the version separator. These new regular expressions use the last @ symbol to be the separator, rather than the first.

Note that in the case of package pinning (that is, "#require PACKAGE@VERSION", the @ is mandatory -- this makes the regex
simpler. In the case of direct usage, we assume the @ is part of the package name if it is the first character.