Restore the ability to require sub-files, but this time with more colons
In some cases, we need to call require() with a specific file in mind, which needs to be encoded as 'package/file'. Building that
into the npm['...'] syntax is tricky, because the PackageCache needs to separate the package names from the files; and packages can
contain slashes in them.

To make this simple to parse, we have introduced a new symbol, ':' to separate the package name from the file to include. Just
before we actually invoke node's require(), we switch it back to a '/'.